refactor(bigquery): reorg query integration tests
