+++ tests/bigquery/src/writes.rs
@@ -14,10 +14,14 @@
 
 mod arrow;
 
-use super::*;
+use super::{INSTANCE_LABEL, random_id_suffix};
 use crate::dataset::{cleanup_stale_datasets, create_dataset, delete_dataset, random_dataset_id};
+use crate::query::UserRecord;
+use anyhow::Result;
+use google_cloud_bigquery::client::BigQuery;
 use google_cloud_bigquery_v2::client::{DatasetService, TableService};
 use google_cloud_bigquery_v2::model::{Table, TableFieldSchema, TableReference, TableSchema};
+use google_cloud_test_utils::runtime_config::project_id;
 
 pub async fn run_writes() -> Result<()> {
     let project_id = project_id()?;
@@ -28,7 +32,7 @@ pub async fn run_writes() -> Result<()> {
     let _ = create_dataset(&dataset_service, &project_id, &dataset_id).await?;
 
     let table_service = TableService::builder().with_tracing().build().await?;
-    let table_id = random_table_id();
+    let table_id = "writes";
 
     let result = async {
         create_table(&table_service, &project_id, &dataset_id, &table_id).await?;
